Add commands for lambda calculus
\lambdaval is boring, \vr and \vl even more so. But ?x is magical: it
now means \vr{x} (that is, \lambdavar{x}, that is, \texttt{x}) in math
mode, while ? in text mode still works as before. Catcodes are awesome.
(And this doesn’t appear to break the older slides, which makes sense
because I can’t come up with a case where I’d want to ask a question in
math mode.)